Update index.md
The below paragraph contains a typo in its first word. It is written "THe" instead of "The"

THe size of the heading text, controlled by the font-size of the <h1> element, has a viewport-relative value with both a maximum and a minimum threshold. To achieve this we use clamp(1.8rem, 2.5vw, 2.8rem). The viewport-relative value is 2.5vw but it is clamped between 1.8rem and 2.8rem, so:

Changing to:
The size of the heading text, controlled by the font-size of the <h1> element, has a viewport-relative value with both a maximum and a minimum threshold. To achieve this we use clamp(1.8rem, 2.5vw, 2.8rem). The viewport-relative value is 2.5vw but it is clamped between 1.8rem and 2.8rem, so: